game: transforms, materials and the filtering call sites
NxMat34 was a 3x3 basis plus a translation; PxTransform is a quaternion
plus a position, so .M/.t become .q/.p and the two places that built a
D3DXMATRIX out of one now go through PxMat33.

Two angle-unit traps, both caught from the 2.8 headers rather than
assumed: NxQuat::fromAngleAxis took DEGREES where PxQuat's angle-axis
constructor takes radians, so the wheel's two quarter turns are PxHalfPi
rather than 90; fromAngleAxisFast already took radians and converts
straight across. Getting this wrong would have rotated wheels by 90
radians and been visible only once the game rendered.

Materials move off NxMaterialDesc onto PxPhysics::createMaterial and are
registered with the index table so shapes keep resolving them.

BEHAVIOUR GAP: anisotropic friction was removed outright in PhysX 3, so
staticFrictionV, dynamicFrictionV, dirOfAnisotropy and NX_MF_ANISOTROPIC
have no equivalent. The original comment says what that was for -- the
car materials were anisotropic specifically so a car sliding along a wall
would skid rather than climb it. Recorded at the site; reproducing it
needs either a contact-modify pass rewriting friction directions or a
wall-specific material.

Also: raycastClosestShape becomes PxScene::raycast with a block buffer,
NxScene::setFilterOps becomes the global PxSetFilterOps, PxShape::getActor
returns a pointer where NxShape::getActor returned a reference,
NxActor::addLocalForce becomes addForce with the force rotated into world
space, and several std::min/std::max calls needed matching types now that
size() is size_t rather than unsigned.

+++ src/Rock3dGame/source/game/FinishMenu.cpp
@@ -96,7 +96,7 @@ void FinishMenu::OnAdjustLayout(const D3DXVECTOR2& vpSize)
 void FinishMenu::OnInvalidate()
 {
 	const Race::Results& results = menu()->GetRace()->GetResults();
-	_playerCount = std::min(results.size(), (unsigned)cBoxCount);
+	_playerCount = static_cast<unsigned>(std::min(results.size(), static_cast<size_t>(cBoxCount)));
 
 	for (unsigned i = 0; i < cBoxCount; ++i)
 	{
